coordinator: Stop raising errors on stopped transactions
When a transaction was gracefully stopped, it indicates that any node
taking part has hit an expected error case, e.g. the primary has
executed hooks which declined a push. The coordinator can't and
shouldn't do anything about it as this is an expected error inside of
the business logic which the Gitaly nodes should know to handle
correctly.

Change the logic to reflect this by silently returning in case the
primary's transaction has been stopped.