Verification of gitaly-ssh runs.
As there is no possibility to check what was passed to gitaly-ssh
process it is hard to say if it works right or not.
To address it and have ability to verify how many times gitaly-ssh
was invoked we include an interception script. The script creates
files with env var and args for each invocation and call the actual
binary just like a proxy. Tests rely on the data present in those
new files to get understanding how many times gitaly-ssh was called
and what parameters were used.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3047