hooks: Check command ported to Go
GitLab-Shell has `bin/check` as a small utility to check the API access
required for the hooks. These hooks are now managed by Gitaly, and
gitaly-hooks has a subcommand `check` too. This subcommend used to
invoke `bin/check` from the vendored `gitlab-shell`.

This change ports the logic to Go, and removes `bin/check`.
It depends on the same config values that the hooks depend on, which
wasn't always true, but usually was just because of Omnibus
configuration reconfiguration.