git2go: Refactor conflict error handling
Due to the process-gap between the conflicts service and its Git2Go
executable, it's not as straight-forward to pass around information as
it usually is. One of those areas which are currently lacking in
conflict handling is how we treat error codes. In case there's a
potentially expected error happening inside the command, it's currently
impossible to tell because all the different errors are lumped together
into a single error return code.

Address this issue by instead serializing error codes into the result
structure. Like this, we can become more informative around the specific
error code and discern different cases.