git: Rename CatFile to ReadObject
The repository interface has recently grown a new function `CatFile`,
which uses git-cat-file(1) to read a Git object from the ODB and returns
its contents. The name is quite implementation-specific though and thus
may be hard to discover for people not familiar with the low-level
plumbing commands. Furthermore, other interfaces like e.g. the
RemoteRepository interface may not use git-cat-file at all to implement
it.

Rename the function to `ReadObject` to make it easier to discover and
less implementation-specific.