operations: Fix PostReceive hook receiving no input
The `updateReferenceWithHooks` function will update a reference and
manually invoke the pre-receive, update and post-receive Git hooks for
the updated reference. To do so, we construct various input parameters
and pass them over to the hook manager, which then executes the custom
hook logic for us. One of these parameters is the standard input
expected by the spawned hooks, which has the same format "<oldrev>
<newrev> <reference>" for both pre-receive and post-receive hooks. As
such, we currently create this buffer once and then pass it to both
functions as a `bytes.Buffer`.

There's a small gotcha here: as the buffer returned by `NewStringBuffer`
is a pointer, it'll get consumed upon reading it when passed to another
function. As a result, as soon as `PostReceiveHook` gets hold of the
buffer it'll be empty already and not contain any references to be
updated. This will lead us to not execute any post-receive hook logic at
all as the assumption now is that no references were updated.

Fix the issue by creating two buffers instead of one only and modify an
existing test case to also check standard input of custom hooks.