port UserCommitFiles to Go
Adds a Go port of UserCommitFiles RPC behind a feature flag. The Go
implementation should mostly match the Ruby implementation's behavior.
Main difference is stricter path validation as the Ruby implementation
did not block every possible path traversal possibility.

To support the RPC, the 'commit' subcommand is added to the gitaly-git2go
executable to build a commit without a worktree. 'commit' uses gob to
serialize the function call parameters to the instead of JSON in order
to preserve type information across the process boundary.