Port UserSquash to Go
Implementation of the UserSquash RPC in Go. The implementation is
hidden behind a feature flag 'go_user_squash'.
The same logic done in Go, not improvements or changes to the logic.
In order to support execution of the commands in a particular directory
SafeBareCmdInDir was added to deal with it.
To support reading from stderr produced by the command a new SafeStderrCmd
was added.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3075