hooks: Gracefully handle lack of config path
If the config file path was not provided to `gitaly-hooks`, it would
panic as the code took the last element of an array which might not have
that size. Now there's a size check, and if the number of elements of
the arguments is too small an error message is displayed.