config: Fix check for executability on systems with strict permissions
Most systems nowadays use a default umask of 0002, which will keep files
to be executable and readable by everybody, but remove the write bit for
others. This is also the assumption made by `config.checkExecutable()`,
which, given a path, checks whether that path is executable or not by
checking that its permission bits are at least 0755. This is wrong in
multiple ways:

    - We don't care for the read and write bits at all.

    - Stricter umasks may remove executable bits. In that case,
      executability depends on the current user and group.

    - Even if a file's mode bits didn't indicate any permissions to
      interact with that file whatsoever, access control lists (ACL) may
      still enable the current user to executable that file.

Fix the check by instead using the access(3P) system call with `X_OK`.
This will match the file's permissions against the current user's creds
any return successfully if the user is allowed to execute that file.