transactions: Remove primary-wins feature flag
In order to ease the migration to reference transactions and reduce the
risk, initially all transactions used a primary-wins strategy: as long
as the primary node performed a vote, the transaction was guaranteed to
succeed. This was thus only a transitional feature flag, and given that
we've been running without for many weeks now without any issues we can
safely remove it now.