hooks: Remove reference-transaction hook feature flag
With the introduction of the Git reference-transaction hook, we
implemented voting on Praefect transaction via this new hook. In order
to keep any fallout of introducing a new hook low, we initially hid that
hook behind a feature flag. And sure enough there had been a bug in
Git's implementation which caused that hook to misbehave. The bug has
since been fixed and no more issues were observed in the last weeks
where it's been running without further incidents in production.

Let's thus remove this feature flag and always enable the new hook. As
there is still an equivalent for the Ruby side, we can't yet remove the
corresponding environment variable.