Fix port UserSquash to Go
This change fixes next issue: when user-squash is used all files in
the target branch disappear after merge and only new files are present.
The cause is that '--no-checkout' flag was set no matter if revision to
checkout was provided into 'worktree add' or not.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3075