gitlab-shell: Test env vars without Ruby exec
This change refactors the `gitlabshell.Env()` test so it doesn't use a
Ruby script to decode the JSON. It depends on the Golang version, and as
such the script can be removed. This is important as it's in the
`ruby/gitlab-shell` directory, and the only script in that whole
directory that still is required in the Gitaly code base.