Add new diff-tree rpc for files changed
Related issue: gitlab.com/gitlab-org/gitlab/-/issues/227572
Related MR: gitlab.com/gitlab-org/gitaly/-/merge_requests/2694

This rpc allows us to get the files changed over a list of commits.
Currently when we are validating a push, we send a CommitDelta for each
commit, which has the potential to be time intensive, and resource
intensive. This rpc allows us to make only one call from Gitlab ->
Gitaly for the same information.