Receiving notifications on changes in database
As we are on the road of adding cache of up to date storages
per repository we need to know when we should invalidate cache
entries in it. In order to archive this we are going to listen
for the changes made to repositories and storage_repositories
tables. The notification contains information about old and new
state of the row. The notification is issued for each changed row.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3053