Extending of git.Repository with Remote interface
In order to support migration of functionality from Ruby to Go implementation
it makes sense to extend existing git.Repository interface with new methods
in order to re-use them during code porting.
This change adds partial implementation of the git remote sub-command.
It is declared as a separate interfaces and could be used separately
from the Repository interface as well. Each command accepts a configuration
struct that allows to set optional flags, etc. Currently it is used only
in tests, but would be in use in the future to omit code duplication.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/1466