proto: Add Tree ID to GitCommit structure
The GitCommit structure doesn't returns or parse the Tree ID, and as
such it's unknown for the clients. It does however help with more
efficient parsing and diffing to know this. For example: Rails currently
can't cache the tree listings based on the Tree ID, but as combination
of reference and path. This is sub optimal as the trees are mostly
shared between commits.

This change also features a slight change regarding to our tests of the
parsing of commits. These were copied throughout, meaning that if we now
add a field to the `GitCommit` structure, the same field has to be added
in a many places. Additionally, having one place with definitions should
make it easier to use it for now tests too.