git: Recognize "vX.Y.GIT" versions
When Git isn't able to detect the version in its build scripts, it will
fall back to use a static version string which is chosen by the
maintainers. This is either the current version (v2.22.0) for a stable
release, a release candidate version (v2.22.0-rc0) or a version
indicating that there's been additional patches without any release
candidate yet (v2.22.GIT). While we correctly parse the former two
variants, the last variant we currently fail to parse.

Recognize the GIT suffix to fix building against master.