ci: also test the 'next' branch in git.git
Improve on a62c6639 ("ci: Add scheduled jobs testing against Git
master", 2020-09-02) by also testing the 'next' branch in git.git.

This'll give us more advanced warning of any breakages, and the 'next'
branch is really quite stable (e.g. Google runs it internally, it's in
Debian experimental etc.).

We could also add "seen" here, but that branch is far more likely to
produce false positives.

Testing this requires pushing a commit which removes the schedule
rule:

    diff --git a/.gitlab-ci.yml b/.gitlab-ci.yml
    index 9940ca11..9302df43 100644
    --- a/.gitlab-ci.yml
    +++ b/.gitlab-ci.yml
    @@ -175,8 +175,6 @@ nightly:git:
           - GO_VERSION: "1.14"
             GIT_VERSION: ["master", "next"]
             TARGET: [ test, test-with-proxies, test-with-praefect ]
    -  rules:
    -    - if: '$CI_PIPELINE_SOURCE == "schedule"'

     cover:
       <<: *cache_definition

As can be seen on
https://gitlab.com/gitlab-org/gitaly/-/pipelines/219137470 it passes
now.

There was an issue where the "next" branch on our mirror wasn't being
updated, but as of writing it's what's in github.com/git/git's "next":
https://gitlab.com/gitlab-org/gitlab-git/-/tree/next

The "seen" branch is still stale from 4 months ago. Maybe someone
kicked or setup a rule for "next", but not "seen":
https://gitlab.com/gitlab-org/gitlab-git/-/tree/seen ?

On naming: This used to be "test:master" before this change, and we
have a "test" target that's testing the matrix of go versions. I
originally went for "test:git_git" (as in git.git) here.

Let's try make it obvious that we're testing various git versions in
the name, but also that this is the nightly CI job, not run in the
regular pipeline. Originally we changed the test target name here, but
now we can keep "nightly:git".