gitaly: Fix deadlock when writing to gRPC streams concurrently
Some of our RPCs directly connect a commands standard streams with gRPC
wrappers to transparently read and write stdin, stdout and stderr.
There's a gotcha though with gRPC. Quoting the documentation of
`grpc.ServerStream`:

	> // It is safe to have a goroutine calling SendMsg and another goroutine
	> // calling RecvMsg on the same stream at the same time, but it is not safe
	> // to call SendMsg on the same stream in different goroutines.
	> SendMsg(m interface{}) error

But we're actually doing exactly that: if a process is writing to
standard output and standard error at the same time, we are now calling
`stream.Send` concurrently. And as `stream.Send` is a direct wrapper of
`SendMsg` we exhibit exactly that kind of behaviour which is explicitly
not allowed.

While we seem to have been lucky in most cases, we did observe odd
behaviour around this in the past with the `ssh.UploadPack` RPC. In some
cases the RPC would spawn a git-upload-pack executable, but it wouldn't
ever exit. This only happened in a very small number of cases though,
which can be attributed to the fact that this behaviour may only be
triggered if the client doesn't support the sideband capability. If it
does, then all of the client's output goes to the multiplexed standard
output, so it wouldn't trigger the issue. But some clients, e.g. old
versions of Git itself or some more obscure implementations, wouldn't
support this capability. If those now trigger a simultaneous write to
stdout and stderr, then the described deadlock happens quite often.

Fix the issue across the tree by using `streamio.NewSyncWriter` instead.
A test is added that invokes the deadlock on my system in about 1/5 of
the testruns.