FetchRemote - porting from ruby to go under feature flag
As we are on the roads of removing ruby code from our codebase
it is a next step in porting of ruby implementations into go code.
FetchRemote RPC now has go implementation that could be enabled
by setting a feature flag 'gitaly_go_fetch_remote'.
To track usage of the new implementation a new counter
'gitaly_fetch_remote_counter' was added (temporary until both
implementations exist).
It also fixes issue with usage of secure token - it should not
be persisted on the disk as well as not present in the list of
the processes with 'ps' command. It has more extended testing
as well.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/1466