git: Add context parameter to `git.Version()`
The `git.Version()` function checks the current Git version which is in
use by Gitaly. To do so, it needs to spawn the git executable. While
it should typically execute and exit quite fast, it's still best
practice to use a context to allow for cancellation of the command.

Add context as a parameter to address this issue.