decouple reconciler from node manager
Reconciler currently depends on NodeManager to perform health checks.
This commit decouples the reconciler from node manager to allow for
alternative health checking implementations. Additionally, this reduces
the surface of the interface and allows reconciler to only depend on the
health information it needs without having to know about the primary.

This decoupling is necessary to hook up the per repository elector stack.