constrain coordinator's node interface
The coordinator only needs to the storage name and the connection
to any given node. Some other components in Praefect, like the replicator,
need full information of the node including the address and the token.
This creates a RouterNode type for the router's and the coordinator to use.
This prevents leaking unneeded information through the interface which would
then force every implementation to pass through.