read-only metrics for repository specific primaries
Adds support for collecting read-only repository metrics taking
per repository primaries in to account. The new functionality is
behind a boolean flag in order to keep metric collection working for
virtual storage primaries as well. In-memory implementations have been
removed as there is no support for per repository primaries in the
local elector.