add a common interface for health checkers
Praefect currently has a separate health checking component to be
used with per repository elector and one integrated in node manager.
Splitting the functionality in a separate component made it easier
to separate the concernrs and test in isolation. This commit adds
a common interface for both and modifies NodeManager to expose the
required methods.