remove concept of primary from replicator
ReplMgr doesn't need to know which node is the primary. Due to
depedning directly on NodeManager, the current interface relies on
the idea of having a primary for a virtual storage. This MR removes
that expectation by decoupling ReplMgr from NodeManager. This is
necessary to hook in the per repository elector as there will be
no primary for a virtual storage anymore.

The interface is replaced with a HealthChecker that simply returns
a list of healthy storages in each virtual storage which are then
used to filter the static set of connections injected as well.

The interface matches what the individual components that build up
the per repository elector provide.