hooks: Pass protocol to `ReceivePackHookEnv()` directly
Hooks expect the GL_PROTOCOL environment variable to be set to the
corresponding protocol that's been used to push to the repository, which
is either "http" for the smarthttp service or "ssh" for the SSH service.
As this is mandatory information for the hook environment, let's move it
into `ReceivePackHookEnv()` directly so it's never missing.