hooks: Remove reftx arguments workaround
The new reference-transaction hook introduced with Git v2.28.0 initially
had an error which caused other hooks to be invoked with arguments of
the reference-transaction hook. To still make it work and use it for
reference transactional voting, we implemented a workaround in c9d0ae5a
(gitaly-hooks: Fixup reference-transaction hook name based on arguments,
2020-08-31) which swapped out the hook name in case we were able to
recognize that faulty behaviour.

The issue has been fixed with Git v2.29.0. Given that Gitaly now
requires that version as the minimum version since a7faf405 (git:
Upgrade minimum required version to v2.29.0, 2020-10-30), we can remove
this workaround.