Update of go-licenses tool
The tool https://github.com/google/go-licenses we use for generating
NOTICE file has some fixes, especially important is one made in
https://github.com/google/go-licenses/pull/19 that add case insensitivity
to regexp to find the license file.
With the new version of the tool we have a fix to our NOTICE file.
The library github.com/git-lfs/go-ntlm/ntlm has a license file named
'License' and it was not handled properly with old versions of
go-licenses tool.