featureset: Simplify description of testcases
Testcases which use the featureset helper currently manually prefix a
"disabled " string to the description everywhere. This is annoying and
doesn't really help, and in fact we can just move it into the feature
set's `String()` method.

Refactor the code by renaming the `String()` method to `Desc()` and
include the prefix.