Resolve dependency on global state in rubyserver.SetHeaders
SetHeaders function relies on the GetObjectDirectoryPath helper.
Internally it access global variable config.Config. This dependency
cause us a lot of issues in testing. To address this issue the code
was moved into Locator abstraction that is now could be injected in
proper places.
As old function GetObjectDirectoryPath removed all places of its
invocation changed to use Locator.GetObjectDirectoryPath.
The func SetHeadersWithoutRepoCheck removed as it not used in prod
code.