operations: Always set GL_PROTOCOL in hooks
When updating references via `updateReferenceWithHooks`, then we need to
assemble the envvars to post to the service. But one envvar actually got
lost in the conversion from Ruby to Go, which is the `GL_PROTOCOL`
variable. Not setting this variable may cause GitLab to misinterpret a
change's source and thus fail to correctly apply access control rules.

Add the missing variable. As the function will only ever be called from
other RPCs in the OperationsService, its safe to assume that it'll
always we "web".