operations: Fix error message when UserMergeToRef conflicts
When creating the merge commit in UserMergeToRef fails with a conflicf,
then we create an error messages "Failed to create merge commit" with
some details filled in. The error message currently confuses the target
SHA and target reference though.

Fix the error message and add a test to verify we got it correct.