git2go: Restrict number of computed virtual merge bases
When doing merges, git uses a three-way merge between three different
states of the file: ours, theirs and the merge base which is the first
common ancestor of ours and theirs. But when both commits have performed
criss-cross merges with each other, then the merge base is not unique.
For example in the following graph:

           o---o <- ours
          / \ /
    base o   X
          \ / \
           o---o <- theirs

When merging ours with theirs, both ours^1 and theirs^1 are immediately
reachable from both commits via ours^2 and theirs^2. As such, both could
serve as a potential merge base for this merge.

What git does in this case is to compute a virtual merge base by merging
ours^1 with theirs^1 first, followed by another merge of this virtual
merge base with ours and theirs. In fact, this is even done recursively
such that if there are repeated criss-cross merges, then git will go all
the way back to the first criss-cross merge and compute all the virtual
merge bases from thereon.

While this is all nice and helps to generate better merges, the result
is that a seemingly simple merge may in fact perform dozens of merges
behind your back. For simple trees this wouldn't hurt much, but as soon
as the tree contains lots of files for which a merge is expensive, then
doing that merge dozens of times quickly starts to hurt as the merge may
now require minutes to compute.

We could solve this issue by restricting recursive merges altogether,
but this would negatively impact the quality of our merges and may
result in more conflicts being generated. Instead, we're now restricting
the number of virtual merge bases being computed. This puts an upper
limit on how many merges are performed and should help avoid abuse via
specially crafted commit histories. If that limit is hit, we simply
pick the first merge base and unravel the remaining virtual merge bases.

This commit picks a kind-of arbitrary limit of 20 computed merge bases.
While picking a lower limit would help performance if we hit such cases,
the downside is that picking a merge base at random may lead to more
conflicts being generated. See e.g. the implemented test case which
nicely demonstrates that the merge base limit does cause us to fail some
merges now.