Fix license permissions to be removable
This fix will ensure that all files added to _build/licenses have
permission 644 for files and 755 for directories. This is needed due to
an issue during notice generation.

The go-licenses tool is used to generate a notice from all of Gitaly's
Go-based dependencies. Sometimes the entire source code is copied as a
notice when the license is unknown. This is the case specifically with
Hashicorp's go-uuid library. When copied, it retains the permissions 444
for files and 555 for directories. This prevents removal when the notice
is regenerated.