add a concurrency safe wrapper for Random
This commit moves Random to it's own file and introduces a wrapper
to make Random safe for concurrent use. The global random in Go
is safe for concurrent use but random instances created otherwise
are not. Adding a locked wrapper allows us to inject a random in to
per repository router and have it provide safe safe access as the
global.