hook up per repository router and elector
This adds a config option to enable the per repository elector and
router. When the `per_repository` elector is configured, the per
repository primary elector and router are enabled and node manager
is disabled.

The node manager is still created and passed as a dependency in
Praefect's info and server serivces as they haven't been updated to
support repository specific primaries yet.

While the configuration option is added, there are still follow ups
to address before the components are safe to use in production. For
example, creating repositories is not supported yet until a follow
up issue is addressed. The changes introduced here are useful for
continued local development repository specific primaries and features
dependent on them such as the variable replication factor.