git.Remote: consistently use {string,bytes}.hasPrefix
Change code added in 57c70475 ("Extending of git.Repository with
Remote interface", 2020-10-23) to use *.hasPrefix consistently. Only
2/3 of the checks of stderr in that commit did, so confusingly when my
9144ba4cf52 ("remote: add meaningful exit code on missing/existing",
2020-10-27) patch landed this check didn't fail on the new version.

Let's make it fail consistently. Why's that a good thing? Because in
the next commit I'll make this use the exit code instead, but now I
won't be as confused because without that patch all 3 tests will fail,
not just 2. This one tolerated the "fatal: MSG" turning into "error:
MSG", not the rest.