git.Remote: use "git remote" exit code instead of parsing stderr
Change the checks added in 57c70475 ("Extending of git.Repository with
Remote interface", 2020-10-23) to use the meaningful exit code I added
upstream in [1].

Luckily we don't need to parse the "git version" output for this, if
it's exiting with 2 or 3 it's the new version, the old version returns
128 and the "fatal: ..." message.

Perhaps I should add an assert() around here somewhere so we'll die if
git.minimumVersion is bumped, but let's just do the same with a
comment for now. We'll eventually run into this again and remove the
dead code sometime after the minimum is bumped to Git v2.30.0.

See [2] for the initial discussion about this code needing an upstream
patch.

To test this I did:

    go test -v ./internal/git/ -run Test.*Remote
    PATH=/home/avar/g/git:$PATH go test -v ./internal/git/ -run Test.*Remote

Where my system git has my upstream [1] patch, and the ~/g/git one
does not (it's a v2.29.0).

1. https://gitlab.com/git-vcs/git/commit/9144ba4cf52
2. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2695#note_436133397