remove automatic reconciler's demo steps
Automatic reconciler has been successfully demoed on multiple
occasions. The only unfilled point is read-only metric missing from
the dashboards. This has an open issue already and is not strictly
related to the automatic reconciliation. This commit thus removes
the automatic reconciliation demo steps to make room for new features
on the demos.