ci: Fix Danger job not running on MRs
With the recent introduction of the MergeRequest pipelines workflow, our
MRs stopped executing the Danger job. The reason is that it's only been
running for branches, which we now won't build anymore due to the
changed workflow.

Fix the issue by only running Danger on merqe requests now, which we
probably should've done in the first place anyway.